+++ packages/fiori-mcp-server/src/page-editor-api/project.ts
@@ -1,13 +1,6 @@
-import {
-    FileName,
-    getCapServiceName,
-    getMinimumUI5Version,
-    getWebappPath,
-    readCapServiceMetadataEdmx
-} from '@sap-ux/project-access';
+import { FileName, getCapServiceName, getMinimumUI5Version, getWebappPath } from '@sap-ux/project-access';
 import type { ApplicationAccess, Manifest, Package } from '@sap-ux/project-access';
 import { FlexChangeLayer } from '@sap/ux-specification/dist/types/src';
-import type { FileData } from '@sap/ux-specification/dist/types/src';
 import { existsSync } from 'node:fs';
 import { readFile } from 'node:fs/promises';
 import { join } from 'node:path';
@@ -49,50 +42,6 @@ export async function getServiceName(appAccess: ApplicationAccess): Promise<stri
         serviceName = await getCapServiceName(appAccess.project.root, appAccess.app?.services?.[serviceName].uri ?? '');
     }
     return serviceName;
-}
-
-/**
- * Reads annotation files for the application.
- *
- * @param appAccess - The application access object
- * @returns A promise that resolves to an array of FileData objects
- */
-export async function readAnnotationFiles(appAccess: ApplicationAccess): Promise<FileData[]> {
-    const annotationData: FileData[] = [];
-    const mainServiceName = getMainService(appAccess);
-    const mainService = appAccess.app?.services?.[mainServiceName];
-    if (!mainService) {
-        return [];
-    }
-    if (mainService.uri && (appAccess.projectType === 'CAPJava' || appAccess.projectType === 'CAPNodejs')) {
-        const serviceUri = mainService?.uri ?? '';
-        if (serviceUri) {
-            const edmx = await readCapServiceMetadataEdmx(appAccess.root, serviceUri);
-            annotationData.push({
-                fileContent: edmx,
-                dataSourceUri: serviceUri
-            });
-        }
-    } else {
-        if (mainService.local) {
-            const serviceFile = await readFile(mainService.local);
-            annotationData.push({
-                dataSourceUri: mainService.local,
-                fileContent: serviceFile.toString()
-            });
-        }
-        const { annotations = [] } = mainService;
-        for (const annotation of annotations) {
-            if (annotation.local) {
-                const annotationFile = await readFile(annotation.local);
-                annotationData.push({
-                    dataSourceUri: annotation.local,
-                    fileContent: annotationFile.toString()
-                });
-            }
-        }
-    }
-    return annotationData;
 }
 
 /**