+++ packages/fiori-app-sub-generator/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/fiori-app-sub-generator
 
+## 0.9.15
+
+### Patch Changes
+
+-   Updated dependencies [7217d7d]
+    -   @sap-ux/project-access@1.32.15
+    -   @sap-ux/annotation-generator@0.3.84
+    -   @sap-ux/axios-extension@1.24.5
+    -   @sap-ux/cap-config-writer@0.12.33
+    -   @sap-ux/fiori-generator-shared@0.13.42
+    -   @sap-ux/launch-config@0.10.42
+    -   @sap-ux/odata-service-inquirer@2.11.20
+    -   @sap-ux/odata-service-writer@0.27.37
+    -   @sap-ux/telemetry@0.6.46
+    -   @sap-ux/ui5-application-inquirer@0.15.46
+    -   @sap-ux/fiori-elements-writer@2.8.18
+    -   @sap-ux/fiori-freestyle-writer@2.5.11
+
 ## 0.9.14
 
 ### Patch Changes