+++ packages/ui5-test-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/ui5-test-writer
 
+## 0.7.42
+
+### Patch Changes
+
+-   Updated dependencies [7217d7d]
+    -   @sap-ux/project-access@1.32.15
+    -   @sap-ux/ui5-application-writer@1.6.3
+
 ## 0.7.41
 
 ### Patch Changes